+++ spec/rubocop/instance_vars_spec.rb
@@ -6,67 +6,57 @@ describe RuboCop::Cop::AnyCable::InstanceVars do
   include_context "cop spec"
 
   it "registers offense for instance var declaration in #subscribed" do
-    inspect_source(<<~RUBY)
+    expect_offense(<<~RUBY)
       class MyChannel < ApplicationCable::Channel
         def subscribed
           @instance_var
+          ^^^^^^^^^^^^^ AnyCable/InstanceVars: Channel instance variables are not supported in AnyCable. Use `state_attr_accessor` instead
         end
       end
     RUBY
-
-    expect(cop.offenses.size).to be(1)
-    expect(cop.messages.first).to include("Channel instance variables are not supported in AnyCable")
   end
 
   it "registers offense for instance var definition inside block in #subscribed" do
-    inspect_source(<<~RUBY)
+    expect_offense(<<~RUBY)
       class MyChannel < ApplicationCable::Channel
         def subscribed
           5.times { @instance_var = 1 }
+                    ^^^^^^^^^^^^^^^^^ AnyCable/InstanceVars: Channel instance variables are not supported in AnyCable. Use `state_attr_accessor` instead
         end
       end
     RUBY
-
-    expect(cop.offenses.size).to be(1)
-    expect(cop.messages.first).to include("Channel instance variables are not supported in AnyCable")
   end
 
   it "registers offense for instance var definition inside condition in #subscribed" do
-    inspect_source(<<~RUBY)
+    expect_offense(<<~RUBY)
       class MyChannel < ApplicationCable::Channel
         def subscribed
           @instance_var = 1 if true
+          ^^^^^^^^^^^^^^^^^ AnyCable/InstanceVars: Channel instance variables are not supported in AnyCable. Use `state_attr_accessor` instead
         end
       end
     RUBY
-
-    expect(cop.offenses.size).to be(1)
-    expect(cop.messages.first).to include("Channel instance variables are not supported in AnyCable")
   end
 
   it "registers offense for instance var definition inside multiple assignments in #subscribed" do
-    inspect_source(<<~RUBY)
+    expect_offense(<<~RUBY)
       class MyChannel < ApplicationCable::Channel
         def subscribed
           a = b = @instance_var = 1
+                  ^^^^^^^^^^^^^^^^^ AnyCable/InstanceVars: Channel instance variables are not supported in AnyCable. Use `state_attr_accessor` instead
         end
       end
     RUBY
-
-    expect(cop.offenses.size).to be(1)
-    expect(cop.messages.first).to include("Channel instance variables are not supported in AnyCable")
   end
 
   it "registers offense for instance var definitions inside action" do
-    inspect_source(<<~RUBY)
+    expect_offense(<<~RUBY)
       class MyChannel < ApplicationCable::Channel
         def follow
           @instance_var = 1
+          ^^^^^^^^^^^^^^^^^ AnyCable/InstanceVars: Channel instance variables are not supported in AnyCable. Use `state_attr_accessor` instead
         end
       end
     RUBY
-
-    expect(cop.offenses.size).to be(1)
-    expect(cop.messages.first).to include("Channel instance variables are not supported in AnyCable")
   end
 end